+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Erl2ex is currently pre-alpha software. Expect significant backwards-incompatibl
 
 *   Convert string() and nonempty_string() Erlang types to the preferred Elixir equivalents, to avoid an Elixir warning.
 *   Support for environment variable interpolation in include paths.
+*   Support for redefining macros.
 
 ## v0.0.5 (2016-01-11)